wip: extract activity instance processing logic to util package
