+++ packages/util/package.json
@@ -9,12 +9,14 @@
   ],
   "dependencies": {
     "@klicker-uzh/prisma": "workspace:*",
-    "@klicker-uzh/types": "workspace:*"
+    "@klicker-uzh/types": "workspace:*",
+    "uuid": "10.0.0"
   },
   "devDependencies": {
     "@rollup/plugin-node-resolve": "~15.3.0",
     "@rollup/plugin-typescript": "~12.1.2",
     "@types/node": "^20.16.1",
+    "@types/uuid": "^10.0.0",
     "cross-env": "~7.0.3",
     "npm-run-all": "~4.1.5",
     "rollup": "~4.34.2",
